pyproject.toml
[project]
name = "benchmork"
version = "0.1.0"
description = "Benchmark your optimization solvers."
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Steven Bytheway" }
]
license = { text = "Apache-2.0" }
keywords = [
    "optimization",
    "solver",
    "benchmarking",
    "linear-programming",
    "operations-research",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Mathematics",
]

dependencies = [
    "numpy>=2.3.5",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "psutil>=7.1.3",
]

[project.optional-dependencies]
# Individual solver dependencies
ortools = ["ortools>=9.7"]
pyomo = ["pyomo>=6.6"]
scipy = ["scipy>=1.11"]
pulp = ["pulp>=3.3.0"]
solvor = ["solvor>=0.3.5"]

# All solvers
all = [
    "ortools>=9.7",
    "pyomo>=6.6",
    "scipy>=1.11",
    "pulp>=3.3.0",
    "solvor>=0.3.5",
]

# Marimo app
app = [
    "marimo>=0.18.4",
    "plotly>=6.5.0",
    "altair>=6.0.0",
]

# Development dependencies
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7",
    "pre-commit>=4.5.1",
]

# Full installation (all solvers + app + dev)
full = [
    "benchmork[all]",
    "benchmork[app]",
    "benchmork[dev]",
]

[project.urls]
Repository = "https://github.com/StevenBtw/benchmORk"

[project.scripts]
benchmork = "benchmork.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["benchmork", "solvers"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["benchmork", "solvers"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "ortools.*",
    "pyomo.*",
    "pulp.*",
    "solvor.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["benchmork", "solvers"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
